backend/requirements.txt
Flask===2.3.2
mysql-connector-python==8.0.33
bcrypt==4.0.1